requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.3
exceptiongroup==1.3.1
fastapi==0.136.1
h11==0.16.0
idna==3.13
passlib==1.7.4
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-multipart==0.0.27
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
